+++ src/gl3.c
@@ -256,7 +256,7 @@ void S2D_GL3_DrawTriangle(GLfloat x1, GLfloat y1,
 /*
  * Draw a texture
  */
-static void S2D_GL3_DrawTexture(double x, double y, double w, double h,
+static void S2D_GL3_DrawTexture(GLfloat x, GLfloat y, GLfloat w, GLfloat h,
                                 GLfloat angle, GLfloat rx, GLfloat ry,
                                 GLfloat r, GLfloat g, GLfloat b, GLfloat a,
                                 GLfloat tx1, GLfloat ty1, GLfloat tx2, GLfloat ty2,
